+++ Sources/Alchemy/Rune/Relationships/Model+Relationships.swift
@@ -1,0 +1,9 @@
+import Foundation
+import NIO
+
+/// Relationships.
+extension Model {
+    public typealias HasOne<To: RelationAllowed> = HasOneRelationship<Self, To>
+    public typealias HasMany<To: RelationAllowed> = HasManyRelationship<Self, To>
+    public typealias BelongsTo<To: RelationAllowed> = BelongsToRelationship<Self, To>
+}
